Add a new File Access preferences pane.
Doing file access on a per-bookmark basis does not work well. When
debugging a local file and a file-line breakpoint were added from the
source view of a stack frame, the user would not be able to then view
the breakpoint in context by selecting it from the breakpoints panel.

Instead, the file access prefs will be used to grant access to directory
subtrees.