Fix loading of large source files.
BSSourceView would wait for the php NSTask to terminate before reading
from the stdout pipe. But if the highlighting output was larger than the
NSPipe buffer, php would fill the buffer and hang in write() waiting for
available space. Instead, start reading the stdout pipe immediately
after launching the task.