signer-ed25519: DER-encode the public and private keys.
The first version of the tool just put the raw key bytes into the PEM
block, which is nonstandard. Use PCKS8 and PKIX to encode the private
and public parts, respectively, which will be compatible with
OpenSSL 3.0. This uses a separate PEM header to disambiguate between the
old form and the new form.